implement combined deleter+test mode
When both -d and -t are passed, run the test
as in test mode against traversed entries, and
delete them only if test find them unproblematic.

While it's a bit hacky, this delivers a mode of
operation where "foreign" symlinks are left intact.
That is, only those entries are removed which are
either produced by reclinker or at least look as if
they were such (that is, they can be restored by a
reclinker invocation).